Added player class to lib folder
