+++ lib/player.rb
@@ -1,0 +1,10 @@
+class Players
+  def initialize
+    @players = []
+  end
+
+  def players(player1, player2)
+    @players << player1
+    @players << player2
+  end
+end
